# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewsFlow API",
        "description": "Documentação oficial da API NewsFlow",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Cadastra um novo usuário",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "responses": {
                    "200": {
                        "description": "Usuário cadastrado com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza login do usuário e retorna token de autenticação",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza logout do usuário autenticado",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Lista todas as categorias",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Categorias listadas com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Cria uma nova categoria",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Exibe uma categoria específica",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "responses": {
                    "200": {
                        "description": "Categoria encontrada com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Atualiza uma categoria existente",
                "operationId": "7dcf71887e585de7f881e9eb98d86126",
                "responses": {
                    "200": {
                        "description": "Categoria atualizada com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Remove uma categoria",
                "operationId": "9b61b65823f4a21f182506b34ddf805c",
                "responses": {
                    "200": {
                        "description": "Categoria removida com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            }
        },
...
        "/login": {
            "get": {
                "tags": [
                    "Sessão Web"
                ],
                "summary": "Exibe a página de login",
                "operationId": "d38b0625602ec8e41c3bb98aa2e969de",
                "responses": {
                    "200": {
                        "description": "Página de login carregada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sessão Web"
                ],
                "summary": "Realiza login via sessão web",
                "operationId": "42f77e072dec84b0e1094cca1228298e",
                "responses": {
                    "302": {
                        "description": "Login realizado e redirecionamento para dashboard"
                    },
                    "422": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
...
        "/register": {
            "get": {
                "tags": [
                    "Registro Web"
                ],
                "summary": "Exibe a página de cadastro de usuário",
                "operationId": "711394cf638654593b284d884e4c5be4",
                "responses": {
                    "200": {
                        "description": "Página de cadastro carregada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Registro Web"
                ],
                "summary": "Realiza o cadastro de um novo usuário via sessão web",
                "operationId": "573de1fed352c1205a32c4d1b9877375",
                "responses": {
                    "302": {
                        "description": "Usuário registrado com sucesso e redirecionado"
                    },
                    "422": {
                        "description": "Dados inválidos ou email já cadastrado"
                    }
                }
            }
        },
...
        "/categories": {
            "get": {
                "tags": [
                    "Categorias Web"
                ],
                "summary": "Lista categorias com contagem de posts",
                "operationId": "ead5fb1cf2a7453acb2e37928a1b1eaf",
                "responses": {
                    "200": {
                        "description": "Lista de categorias exibida com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorias Web"
                ],
                "summary": "Cria uma nova categoria",
                "operationId": "bb02f0320882db37f99574196577e446",
                "responses": {
                    "302": {
                        "description": "Categoria criada com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
...
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categorias Web"
                ],
                "summary": "Exibe uma categoria específica",
                "operationId": "4cdf41756315109f1efc72787db5830a",
                "responses": {
                    "200": {
                        "description": "Categoria exibida com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorias Web"
                ],
                "summary": "Atualiza uma categoria existente",
                "operationId": "19954b1f8aa4916b5f032dc28ea36a24",
                "responses": {
                    "302": {
                        "description": "Categoria atualizada com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorias Web"
                ],
                "summary": "Remove uma categoria",
                "operationId": "f1b8af6e84bd52b9594d98bde748ebd6",
                "responses": {
                    "302": {
                        "description": "Categoria deletada com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                }
            }
        },
...
    },
...
}